blog-system/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
python-decouple>=3.8
django-filter>=23.3
markdown>=3.5.0
bleach>=6.1.0

